Moved helper functions, unified display controller
