+++ src/bottles.py
@@ -6,31 +6,16 @@ This includes all functions for the Lists, DB and Buttos/Dropdowns.
 from config.config_manager import shared
 from src.error_suppression import logerror
 
-from src.display_handler import DisplayHandler
 from src.database_commander import DatabaseCommander
 from src.display_controller import DisplayController
 from src.rpi_controller import RpiController
 from src.logger_handler import LoggerHandler
-from src.supporter import (
-    generate_CBB_names,
-    generate_LBelegung_names,
-    generate_PBneu_names,
-    generate_ProBBelegung_names,
-)
 
 
-DP_HANDLER = DisplayHandler()
 DB_COMMANDER = DatabaseCommander()
 DP_CONTROLLER = DisplayController()
 RPI_CONTROLLER = RpiController()
 LOG_HANDLER = LoggerHandler("bottles_module", "production_logs")
-
-
-def customlevels(w):
-    """ Opens the additional window to change the volume levels of the bottles. """
-    bot_names = []
-    vol_values = []
-    w.bottleswindow(bot_names, vol_values)
 
 
 def get_bottle_ingredients():
@@ -42,7 +27,7 @@ def get_bottle_ingredients():
 def refresh_bottle_cb(w):
     """ Adds or remove items to the bottle comboboxes depending on the changed value"""
     # Creating a list of the new and old bottles used
-    combobox_bottles = generate_CBB_names(w)
+    combobox_bottles = DP_CONTROLLER.get_comboboxes_bottles(w)
     old_order = shared.old_ingredient
     new_order = DP_CONTROLLER.get_current_combobox_items(combobox_bottles)
 
@@ -51,7 +36,7 @@ def refresh_bottle_cb(w):
     new_bottle = new_blist[0] if new_blist else ""
     old_bottle = old_blist[0] if old_blist else ""
 
-    DP_HANDLER.adjust_bottle_comboboxes(combobox_bottles, old_bottle, new_bottle)
+    DP_CONTROLLER.adjust_bottle_comboboxes(combobox_bottles, old_bottle, new_bottle)
 
     register_bottles(w)
     shared.old_ingredient = new_order
@@ -61,7 +46,7 @@ def refresh_bottle_cb(w):
 def calculate_combobox_bottles(w):
     """ Fills each bottle combobox with the possible remaining options
     """
-    combobox_bottles = generate_CBB_names(w)
+    combobox_bottles = DP_CONTROLLER.get_comboboxes_bottles(w)
     used_ingredients = shared.old_ingredient
     possible_ingredients = DB_COMMANDER.get_ingredient_names_machine()
 
@@ -70,7 +55,7 @@ def calculate_combobox_bottles(w):
         used_without_self = {x for i, x in enumerate(used_ingredients) if i != row}
         shown_ingredients.append(sorted(set(possible_ingredients) - used_without_self))
 
-    DP_HANDLER.fill_multiple_combobox_individually(combobox_bottles, shown_ingredients, True)
+    DP_CONTROLLER.fill_multiple_combobox_individually(combobox_bottles, shown_ingredients, True)
 
 
 @logerror
@@ -80,7 +65,7 @@ def register_bottles(w):
     from src.maker import refresh_recipe_maker_view
 
     # Checks where are entries and appends them to a list
-    combobox_bottles = generate_CBB_names(w)
+    combobox_bottles = DP_CONTROLLER.get_comboboxes_bottles(w)
     ingredient_names = DP_CONTROLLER.get_current_combobox_items(combobox_bottles)
     DB_COMMANDER.set_bottleorder(ingredient_names)
 
@@ -93,47 +78,46 @@ def register_bottles(w):
 @logerror
 def read_in_bottles(w):
     """ Reads the Bottleorder into the BottleTab. """
-    combobox_bottles = generate_CBB_names(w)
+    combobox_bottles = DP_CONTROLLER.get_comboboxes_bottles(w)
     ingredient_names = DB_COMMANDER.get_ingredients_at_bottles()
-    DP_HANDLER.set_multiple_combobox_items(combobox_bottles, ingredient_names)
+    DP_CONTROLLER.set_multiple_combobox_items(combobox_bottles, ingredient_names)
 
 
 @logerror
 def refresh_bottle_information(w):
     """ Loads or updates the Labels of the Bottles (Volumelevel). """
-    labels = generate_LBelegung_names(w)
     label_names = DB_COMMANDER.get_ingredients_at_bottles()
     label_names = [f"  {x}:" if x != "" else "  -  " for x in label_names]
-    DP_HANDLER.fill_multiple_lineedit(labels, label_names)
+    DP_CONTROLLER.set_label_bottles(w, label_names)
 
 
 @logerror
 def renew_checked_bottles(w):
     """ Renews all the Bottles which are checked as new. """
-    pushbutton_new_list = generate_PBneu_names(w)
+    pushbutton_new_list = DP_CONTROLLER.get_pushbottons_newbottle(w)
     renew_bottle = DP_CONTROLLER.get_toggle_status(pushbutton_new_list)
     DB_COMMANDER.set_bottle_volumelevel_to_max(renew_bottle)
-    DP_HANDLER.untoggle_buttons(pushbutton_new_list)
+    DP_CONTROLLER.untoggle_buttons(pushbutton_new_list)
     set_fill_level_bars(w)
-    DP_HANDLER.standard_box("Alle Flaschen angewendet!")
+    DP_CONTROLLER.standard_box("Alle Flaschen angewendet!")
 
 
 @logerror
 def set_fill_level_bars(w):
     """ Gets the proportion of actual and maximal volume of each connected bottle and asigns it"""
-    progressbars = generate_ProBBelegung_names(w)
+    progressbars = DP_CONTROLLER.get_levelbar_bottles(w)
     fill_levels = DB_COMMANDER.get_bottle_fill_levels()
-    DP_HANDLER.set_progress_bar_values(progressbars, fill_levels)
+    DP_CONTROLLER.set_progress_bar_values(progressbars, fill_levels)
 
 
 @logerror
 def clean_machine(w):
     """ Activate all Pumps for 20 s to clean them. Needs the Password. Logs the Event. """
     if not DP_CONTROLLER.check_bottles_password(w):
-        DP_HANDLER.standard_box("Falsches Passwort!!!!")
+        DP_CONTROLLER.standard_box("Falsches Passwort!!!!")
         return
 
-    DP_HANDLER.standard_box("Achtung!: Maschine wird gereinigt, genug Wasser bereitstellen! Ok zum Fortfahren.")
+    DP_CONTROLLER.standard_box("Achtung!: Maschine wird gereinigt, genug Wasser bereitstellen! Ok zum Fortfahren.")
     LOG_HANDLER.log_header("INFO", "Cleaning the Pumps")
     RPI_CONTROLLER.clean_pumps()
-    DP_HANDLER.standard_box("Fertig!!!")
+    DP_CONTROLLER.standard_box("Fertig!!!")
